Updated README, installRaspbian and installApplication script.

+++ installApplication.sh
@@ -1,19 +1,29 @@
 #!/bin/bash
 
-sudo apt-get update
-sudo apt-get upgrade
+set -e
 
-sudo apt-get install \
-  curl \
-  wiringpi \
-  libsqlite3-dev \
-  build-essential
+if [[ ! -f ~/initial.lock ]]
+  # initialize locales
+  sudo /usr/sbin/locale-gen
 
-curl -L -o sprinklers_pi.tar.gz https://github.com/rszimm/sprinklers_pi/archive/v1.5.3.tar.gz
-mv sprinklers_pi-* sprinklers_pi
+  # initial update and upgrade on first boot
+  sudo apt-get update
+  sudo DEBIAN_FRONTEND=noninteractive apt-get -y upgrade
 
-make
-sudo make install
+  # install mandatory resources
+  sudo apt-get -y install \
+    curl \
+    wiringpi \
+    libsqlite3-dev \
+    build-essential
 
+  # download and install latest sprinklers_pi version
+  curl -L -o sprinklers_pi.tar.gz https://github.com/rszimm/sprinklers_pi/archive/v1.5.3.tar.gz
+  mv sprinklers_pi-* sprinklers_pi
+  cd sprinklers_pi
+  make
+  sudo make install
 
-sudo reboot now
+  touch ~/initial.lock
+  sudo reboot now
+fi
